// src/app/admin/models/add-new-template.model.ts
export class AddNewTemplatePropertyModel {
    propertyId: number;
    companyId: number;
    requestTypeId: number;
    propertyName: string;
    propertyType: string;
    apiPropertyName: string;
    parentId: number;
    isRequired: boolean;
    isVisibleOnUI: boolean;
    length: number;
    orderBy: number;
}